1. Move ORDER_CA into cli as environmental variable 2. Make ordering policy as both Org1 and Org2 need to endorse. But seemed invoke transaction successful, yet value not updated. 3. TODO: Add fabric-explore 4. TODO: Have User1 invoke and query transaction.

+++ stopFabricNetwork.sh
@@ -16,7 +16,10 @@ VERSION=1.1.0
 BASE_DOCKER_TAG=x86_64-0.4.9
 
 function teardownFabricNetwork(){
-    docker-compose -f ${FABRIC_ROOT}/docker-compose-e2e.yaml down
+    file=docker-compose-e2e.yaml
+    if [ -f $file ]; then
+    	docker-compose -f ${FABRIC_ROOT}/docker-compose-e2e.yaml down
+    fi
 
     CONTAINER_IDS=$(docker ps -aq)
     if [ -z "$CONTAINER_IDS" -o "$CONTAINER_IDS" = " " ]; then
